v1/requirements.txt
aioredis
aiohttp
redis==5.2.1
python-docx
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
aiocache
ollama
litellm
openai>=1.67.0
httpx
python-dotenv
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest==7.4.4
pytest-asyncio==0.23.3
psutil==6.1.0  # For system monitoring
dspy